float形式的配置需要用memcpy方式从共享内存拷入拷出

+++ inc/StateManger.h
@@ -19,7 +19,7 @@ public:
 	~StateManger();
 
 	void* returnIpc();
-	void GetParams(float* p);
+	void GetParams(int* p);
 	int CurStateInterface();
 
 	void ToStateAuto_Mtd();
